requirements.txt
discord.py>=2.0
aiohttp>=3.8
python-dotenv>=1.0
pillow>=10.0
pillow-heif>=0.10
pytest>=7.0
